Adds CardAdapter, new card layouts, and ViewPager2 setup
This commit introduces a `CardAdapter` and three new layout files (`link_check_card`, `browser_pick_card`, and `settings_card`) to represent different content cards.

The `MainActivity` is updated to use `ViewPager2` to display these cards with a visual effect that shows the next card peeking. The background color of the main layout is also changed to a shade of purple. ConstraintLayout and CardView dependencies are added to the build file, along with ViewPager2. A new `dimens.xml` file is added for ViewPager2 padding.